Move docs module to core
