+++ kartothek/core/dataset.py
@@ -15,6 +15,7 @@ from kartothek.core import naming
 from kartothek.core._compat import load_json
 from kartothek.core._mixins import CopyMixin
 from kartothek.core.common_metadata import SchemaWrapper, read_schema_metadata
+from kartothek.core.docs import default_docs
 from kartothek.core.index import (
     ExplicitSecondaryIndex,
     IndexBase,
@@ -349,6 +350,7 @@ class DatasetMetadataBase(CopyMixin):
         combined_indices.update(indices)
         return self.copy(indices=combined_indices)
 
+    @default_docs
     def get_indices_as_dataframe(
         self,
         columns: Optional[List[str]] = None,
@@ -370,11 +372,6 @@ class DatasetMetadataBase(CopyMixin):
 
         Parameters
         ----------
-        columns: list of str
-            If provided, the dataframe will only be constructed for the provided columns/indices.
-            If `None` is given, all indices are included.
-        date_as_object: bool, optional
-            Cast dates to objects.
         """
         if columns is None:
             columns = sorted(self.indices.keys())
