+++ kartothek/io/dask/dataframe.py
@@ -6,9 +6,9 @@ import dask.dataframe as dd
 import numpy as np
 
 from kartothek.core.common_metadata import empty_dataframe_from_schema
+from kartothek.core.docs import default_docs
 from kartothek.core.factory import _ensure_factory
 from kartothek.core.naming import DEFAULT_METADATA_VERSION
-from kartothek.io_components.docs import default_docs
 from kartothek.io_components.metapartition import parse_input_to_metapartition
 from kartothek.io_components.update import update_dataset_from_partitions
 from kartothek.io_components.utils import (
